backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic-settings==2.8.0
SQLAlchemy==2.0.38
alembic==1.14.1
psycopg[binary]==3.2.13
redis==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.20
email-validator==2.2.0
